+++ NEWS.md
@@ -1,5 +1,8 @@
 # vaster dev
 
+* New function `vaster_listxyz()` to convert extent, dimension, and optionally data to trad `image()` format. 
+
+
 * New function `vaster_long()` to convert extent, dimension, and optionally data to long form XYZ in raster or trad matrix order. 
 
 * All obvious raster functions now included (needs more testing).